javascript DOM for the style attribute, .css file format output
This update duplicates how the style attribute is accessed when writing
javascript. This is a nice feature to have for the people that are
familiar with it. While I was creating this I thought that it would
also be handy to have it output in .css format as well. So I added that
ability also.