html_tag: create hyphenate_tagname attribute
Python does not allow classes to be created with hyphenated names, and
it's convenient to be able to use the tag's Python class name for
rendering, rather than specifying an explicit tagname attribute for each
custom tag.

HTML allows hyphenated tag names, so add an attribute to html_tag to
replace underscores in the tag's Python class name with hyphens when enabled.